fix: migrate firebase configuration to firebaseConfig.js and implement persistence handling
