+++ firebaseConfig.js
@@ -1,0 +1,31 @@
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import { initializeApp } from "firebase/app";
+import { getReactNativePersistence, initializeAuth } from "firebase/auth";
+import { getFirestore } from "firebase/firestore";
+
+// Configuración de Firebase
+const firebaseConfig = {
+  apiKey: process.env.EXPO_PUBLIC_FIREBASE_API_KEY,
+  authDomain: process.env.EXPO_PUBLIC_FIREBASE_AUTH_DOMAIN,
+  projectId: process.env.EXPO_PUBLIC_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.EXPO_PUBLIC_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.EXPO_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.EXPO_PUBLIC_FIREBASE_APP_ID,
+};
+
+// Inicializar App
+const app = initializeApp(firebaseConfig);
+
+/**
+ * SOLUCIÓN AL ERROR 'setItem':
+ * Creamos un objeto de persistencia que asegura la compatibilidad total.
+ * Firebase v10+ a veces falla al detectar los métodos de AsyncStorage directamente.
+ */
+const auth = initializeAuth(app, {
+  persistence: getReactNativePersistence(AsyncStorage),
+});
+
+// Inicializar Firestore
+const db = getFirestore(app);
+
+export { auth, db };
